requirements.txt
https://github.com/Sam-Joshua-S/Install-dlib/blob/main/dlib-19.19.0-cp37-cp37m-win_amd64.whl
albumentations==1.3.0
datasouls_antispoof==0.0.4
face_recognition==1.3.0
Flask==2.2.3
flask_login==0.6.2
flask_sqlalchemy==3.0.3
mediapipe==0.9.0.1
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
bcrypt==4.0.1
pyttsx3==2.90
SQLAlchemy==2.0.8
torch==1.13.1
torchvision==0.14.1
iglovikov-helper-functions